+++ packages/color-picker/src/pickers/TailwindColorPicker.tsx
@@ -1,22 +1,26 @@
-import colors from 'tailwindcss/colors';
 import { ColorGrid } from '~/components/ColorGrid';
 import type { StringBasedPickerProps } from '~/types';
+import tw4Colors from '~/data/tw4-colors.json';
+import { parseColor, formatSchemistTo } from '@palettebro/theme-generator';
 
 export const TailwindColorPicker: React.FC<StringBasedPickerProps> = ({
   onChange,
 }: {
   onChange: (v: string) => void;
 }) => {
-  const skipColors = ['current', 'transparent', 'black', 'white', 'inherit'];
-  const options = Object.keys(colors)
-    .filter((color) => !skipColors.includes(color))
-    .map((color) => ({
-      id: color,
-      color: colors[color as keyof typeof colors][500],
-      borderColor: colors[color as keyof typeof colors][600],
-      label: colors[color as keyof typeof colors][500],
-      sublabel: color,
-    }));
+  const handleChange = (color: string) => {
+    const [colorFormat, schemistColor] = parseColor(color);
+    if (!schemistColor || !colorFormat) return;
+    onChange(formatSchemistTo(schemistColor, 'hex'));
+  };
 
-  return <ColorGrid options={options} onChange={onChange} />;
+  const options = Object.entries(tw4Colors).map(([color, value]) => ({
+    id: color,
+    color: value,
+    borderColor: value,
+    label: color,
+    sublabel: color,
+  }));
+
+  return <ColorGrid options={options} onChange={handleChange} />;
 };