+++ packages/color-picker/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @palettebro/color-picker
 
+## 0.2.0
+
+### Minor Changes
+
+- Remove TW colors dep
+
+### Patch Changes
+
+- Updated dependencies
+  - @palettebro/theme-generator@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes