+++ packages/tailwind-theme/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @palettebro/tailwind-theme
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @palettebro/theme-generator@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes