+++ packages/theme-generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @palettebro/theme-generator
 
+## 0.2.0
+
+### Minor Changes
+
+- Remove TW colors dep
+
 ## 0.1.0
 
 ### Minor Changes